requirements.txt
Flask==2.3.2
Streamlit
opencv-python
numpy
face-recognition
dlib
pygame
ultralytics
folium
pandas
matplotlib
streamlit_folium
insightface
onnxruntime
pillow
requests
torch
scikit-learn
plotly